Changed pin C9 to default input instead of MC02
When the RNH powers up we want the KSZ in a full off state including
no clocks. When the pin is in MC02 mode it outputs some (selectable)
clock. There is no way to shut it off as far as I can tell short of
setting the pin to a different mode.