Adis project correctly sends data now
It's ugly and I hate it and I'm going to change it. It turns out
the problem was the addresses for everything ever were all wrong, and
consequently are next on the chopping block to redo ~correctly~.
Also of useful note is we should be doing dev with -O0 only, higher
optimization levels end up removing some asserts.