One more refactor of net_addrs/utils_sockets
I'm finally happy with the public facing side of these, this is
likely net_addrs final form. It's not identical to what is in
av3fc, but we've decided to make the canonical form a submodule
of all projects needing it, and then have a code generator spit out
the best implementation for each project.