Replace servo filter with explicit controls.
The filter was hard to reason about and also couldn't guarantee servo
safety under certain pathological inputs (such as n hard lefts followed
by n hard rights, where n is the number of samples in the moving average
filter).

Now the servo's behavior is explicitly regulated such that:
  1) its speed is explicity regulated to be below a compile-time limit;
  2) the frequency at which it changes direction is also explicitly
     regulated to be below a different compile time limit.
If the commands given to the servo would cause its speed to be too high,
instead a new output is chosen using the maximum safe servo speed.
If the commands given to the servo would cause it to change direction
too quickly, it instead keeps the servo at its current position until
either enough time has elapsed that it is allowed to change direction
again, or a new command that does not change the servos direction is
received.