Add empirically-determined servo limits.
These limits have been determined by Dan Lidral-Porter, K Wilson, and
Dave Camarillo at the PSAS meeting on 3/25/13.

We were unsure of what the full intended range of movement was (since
the 15 degrees from center that Dave quoted was clearly too far), and we
found that there is a mechanical asymmetry in the linkage. So we
eyeballed some limits based on getting close to making the linkage
points collinear. These limits should be revisited once the aeorodynamic
requirements have been sussed out.