Disabled RCI startup lwip check
Since lwip starts in a different thread, even though it was started
before the call to RCICreate, sometimes that thread wasn't making it
to where it set its own name before RCICreate checked for it.

Sometimes it was going so far as to make it to the point in the RCI
thread that gets a socket, causing it to fail. A stopgap measure of
sleeping 100ms seems to fix the problem while a better solution is
searched for.