Fixed Port current ADC measurements bleeding into following ones
We had been switching the analog mux and then immediately beginning
an ADC conversion, but this resulted in measurements bleeding over
into the following ones. Now it swaps the mux half way between one
conversion finishing and another starting.

Ideally It'd swap as soon as a conversion completes, but coordinating
between the two ADCs is tricky. I may fix this later.