More attempts to get KSZ8081RND phy working
This might do it, but I don't have a cable to test with.

A few things going on here, the mac_lld driver has a BOARD_PHY_RESET
macro that we can hook our own reset in with, which is good because
the 8081 D variant needs a register set from using a 50MHz clock to
a 25MHz clock at reset (the A variant assumes 25MHz at reset). There's
also other junk in there like setting the NRST line and doing a SW
reset.

Defining this correctly gets us through mac_lld_init, but when
mac_lld_start is called (as part of lwipthread, that's why that is
all there) it hangs on polling ETH->DMAMBR for software reset.
According to the internets that can hang if the MII ref clock from the
phy to the chip is off. It's off because at the end of mac_lld_init
the phy gets turned off, which apparently resets the phy input clock
select register. It would be great if there was something like
BOARD_PHY_RESET at the beginning of mac_lld_start, but in lieu of that
we can just stop the mac driver from turning the phy off by setting
STM32_ETH1_CHANGE_PHY_STATE to FALSE in mcuconf.h