Updated MAX2769 to work with GPSv3
The biggest change here is breaking the config out into two structs, one
for the CPLD and one for the MAX. This lets me simplify a bunch of the
names.

Some prelim CPLD setup is done, but it's not correct at the moment,
hence why the last two lines are commented out in MAX279_start()

Also removed some unused debug macros, and renamed the data ready event.