Fixed i2c hitting an assert after timeout
On a timeout ChibiOS will set the bus into I2C_LOCKED, which will
cause an assert to be hit if a bus transfer is tried again.
I2C_LOCKED can only be cleared by i2cStart(), but i2cStart will hit
an assert if i2cStop is not issued before restarting.

This seems like a really terrible way to handle timeous