Remove timeout support from venus serial transfers.
Initially we wanted to cause all venus network sends to be initiated by
a timer, but stopping a uart transfer was too slow due to DMA, and was
missing bytes here and there. Next we wanted a timer to sync the uart
transfers to a venus message, but chibiOS gives us no good hooks to id
when a character arrives. So unfortunately we must remove timer code.